+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 ### is.codion.framework.domain
 - DefaultEntityValidator removed, validation moved to AbstractAttributeDefinition, EntityValidator now with default method implementations.
 - Attribute.Type.isNumerical() renamed to isNumeric().
+- DefaultColumnDefinition, Column.SetParameter instances now shared.
 ### is.codion.framework.db
 - DefaultSelect bug fixed, having clause not included in hashCode().
 - Select.having() added, related refactoring.